requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
redis==5.0.7
aiofiles==23.2.1
openai==1.35.10
pyyaml==6.0.1
tenacity==8.4.1
python-json-logger==2.0.7
httpx==0.27.2
